@@ -171,97 +171,96 @@
     ]
   },
   "devDependencies": {
-    "@babel/core": "^7.20.5",
-    "@babel/plugin-transform-runtime": "^7.19.6",
-    "@babel/preset-env": "^7.20.2",
+    "@babel/core": "^7.21.8",
+    "@babel/plugin-transform-runtime": "^7.21.4",
+    "@babel/preset-env": "^7.21.5",
     "@babel/preset-react": "^7.18.6",
-    "@babel/preset-typescript": "^7.18.6",
-    "@graphql-codegen/cli": "2.16.1",
-    "@graphql-codegen/fragment-matcher": "3.3.3",
-    "@graphql-codegen/typescript": "2.8.5",
-    "@graphql-codegen/typescript-operations": "2.5.10",
+    "@babel/preset-typescript": "^7.21.5",
+    "@graphql-codegen/cli": "3.3.1",
+    "@graphql-codegen/fragment-matcher": "4.0.1",
+    "@graphql-codegen/typescript": "3.0.4",
+    "@graphql-codegen/typescript-operations": "3.0.4",
     "@graphql-codegen/typescript-react-apollo": "3.3.7",
     "@pmmmwh/react-refresh-webpack-plugin": "0.5.10",
     "@simbathesailor/use-what-changed": "^2.0.0",
-    "@teamsupercell/typings-for-css-modules-loader": "^2.5.1",
+    "@teamsupercell/typings-for-css-modules-loader": "^2.5.2",
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@types/jest": "^29.2.4",
-    "@types/lodash": "^4.14.191",
+    "@testing-library/react": "^14.0.0",
+    "@types/jest": "^29.5.1",
+    "@types/lodash": "^4.14.194",
     "@types/mousetrap": "^1.6.11",
-    "@types/node": "18.11.13",
-    "@types/react": "^18.0.26",
-    "@types/react-dom": "^18.0.9",
+    "@types/node": "20.2.1",
+    "@types/react": "^18.2.6",
+    "@types/react-dom": "^18.2.4",
     "@types/react-test-renderer": "^18.0.0",
     "@types/redux-logger": "^3.0.9",
     "@types/styled-components": "^5.1.26",
     "@types/terser-webpack-plugin": "^5.0.4",
     "@types/webpack-env": "^1.18.0",
-    "@typescript-eslint/eslint-plugin": "^5.46.0",
-    "@typescript-eslint/parser": "^5.46.0",
-    "babel-loader": "^9.1.0",
+    "@typescript-eslint/eslint-plugin": "^5.59.6",
+    "@typescript-eslint/parser": "^5.59.6",
+    "babel-loader": "^9.1.2",
     "babel-plugin-direct-import": "^1.0.0",
-    "browserslist": "^4.21.4",
-    "chalk": "^4.1.2",
-    "concurrently": "^7.6.0",
+    "browserslist": "^4.21.5",
+    "concurrently": "^8.0.1",
     "copy-webpack-plugin": "^11.0.0",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.7.2",
-    "css-minimizer-webpack-plugin": "^4.2.2",
+    "css-loader": "^6.7.3",
+    "css-minimizer-webpack-plugin": "^5.0.0",
     "detect-port": "^1.5.1",
-    "electron": "^24.1.0",
-    "electron-builder": "^23.6.0",
+    "electron": "^24.3.1",
+    "electron-builder": "^24.4.0",
     "electron-devtools-installer": "^3.2.0",
     "electron-notarize": "^1.1.1",
     "electron-rebuild": "^3.2.5",
-    "eslint": "^8.29.0",
+    "eslint": "^8.40.0",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-config-erb": "^4.0.3",
-    "eslint-import-resolver-typescript": "^3.5.2",
+    "eslint-config-erb": "^4.0.6",
+    "eslint-import-resolver-typescript": "^3.5.5",
     "eslint-import-resolver-webpack": "^0.13.2",
-    "eslint-plugin-compat": "^4.0.2",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^27.1.6",
-    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-compat": "^4.1.4",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jest": "^27.2.1",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-promise": "^6.1.1",
-    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.5.0",
-    "husky": "^8.0.2",
+    "html-webpack-plugin": "^5.5.1",
+    "husky": "^8.0.3",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^29.3.1",
-    "lint-staged": "^13.1.0",
-    "mini-css-extract-plugin": "^2.7.2",
-    "prettier": "^2.8.1",
+    "jest": "^29.5.0",
+    "lint-staged": "^13.2.2",
+    "mini-css-extract-plugin": "^2.7.5",
+    "prettier": "^2.8.8",
     "react-refresh": "^0.14.0",
-    "react-refresh-typescript": "^2.0.7",
+    "react-refresh-typescript": "^2.0.9",
     "react-test-renderer": "^18.2.0",
-    "rimraf": "^3.0.2",
-    "sass": "^1.56.2",
-    "sass-loader": "^13.2.0",
+    "rimraf": "^5.0.1",
+    "sass": "^1.62.1",
+    "sass-loader": "^13.2.2",
     "source-map-loader": "^4.0.1",
-    "style-loader": "^3.3.1",
-    "terser-webpack-plugin": "^5.3.6",
-    "ts-jest": "^29.0.3",
+    "style-loader": "^3.3.2",
+    "terser-webpack-plugin": "^5.3.9",
+    "ts-jest": "^29.1.0",
     "ts-loader": "^9.4.2",
     "ts-node": "^10.9.1",
     "url-loader": "^4.1.1",
-    "webpack": "^5.75.0",
-    "webpack-bundle-analyzer": "^4.7.0",
-    "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^4.11.1",
+    "webpack": "^5.83.1",
+    "webpack-bundle-analyzer": "^4.8.0",
+    "webpack-cli": "^5.1.1",
+    "webpack-dev-server": "^4.15.0",
     "webpack-font-preload-plugin": "^1.5.0",
     "webpack-merge": "^5.8.0",
     "workbox-webpack-plugin": "^6.5.4"
   },
   "dependencies": {
-    "@apollo/client": "^3.7.2",
+    "@apollo/client": "^3.7.14",
     "@exponent/electron-cookies": "^2.0.0",
-    "@mui/icons-material": "^5.10.16",
-    "@mui/lab": "^5.0.0-alpha.111",
-    "@mui/material": "^5.10.17",
-    "@mui/styled-engine-sc": "^5.10.16",
+    "@mui/icons-material": "^5.11.16",
+    "@mui/lab": "^5.0.0-alpha.130",
+    "@mui/material": "^5.13.1",
+    "@mui/styled-engine-sc": "^5.12.0",
     "@structure/common": "file:../common",
     "@structure/config": "file:../config",
     "apollo3-cache-persist": "^0.14.1",
