@@ -1,7 +1,6 @@
 // Check if the renderer and main bundles are built
-import path from 'path';
-import chalk from 'chalk';
 import fs from 'fs';
+import path from 'path';
 import webpackPaths from '../configs/webpack.paths';
 
 const mainPath = path.join(webpackPaths.distMainPath, 'main.js');
@@ -9,16 +8,12 @@ const rendererPath = path.join(webpackPaths.distRendererPath, 'renderer.js');
 
 if (!fs.existsSync(mainPath)) {
   throw new Error(
-    chalk.whiteBright.bgRed.bold(
-      'The main process is not built yet. Build it by running "npm run build:main"'
-    )
+    'The main process is not built yet. Build it by running "npm run build:main"'
   );
 }
 
 if (!fs.existsSync(rendererPath)) {
   throw new Error(
-    chalk.whiteBright.bgRed.bold(
-      'The renderer process is not built yet. Build it by running "npm run build:renderer"'
-    )
+    'The renderer process is not built yet. Build it by running "npm run build:renderer"'
   );
 }
