@@ -1,6 +1,5 @@
-import fs from 'fs';
-import chalk from 'chalk';
 import { execSync } from 'child_process';
+import fs from 'fs';
 import { dependencies } from '../../package.json';
 
 if (dependencies) {
@@ -25,24 +24,19 @@ if (dependencies) {
     if (filteredRootDependencies.length > 0) {
       const plural = filteredRootDependencies.length > 1;
       console.log(`
- ${chalk.whiteBright.bgYellow.bold(
-   'Webpack does not work with native dependencies.'
- )}
-${chalk.bold(filteredRootDependencies.join(', '))} ${
+ Webpack does not work with native dependencies.
+
+ ${filteredRootDependencies.join(', ')} ${
         plural ? 'are native dependencies' : 'is a native dependency'
       } and should be installed inside of the "./release/app" folder.
  First, uninstall the packages from "./package.json":
-${chalk.whiteBright.bgGreen.bold('npm uninstall your-package')}
- ${chalk.bold(
-   'Then, instead of installing the package to the root "./package.json":'
- )}
-${chalk.whiteBright.bgRed.bold('npm install your-package')}
- ${chalk.bold('Install the package to "./release/app/package.json"')}
-${chalk.whiteBright.bgGreen.bold('cd ./release/app && npm install your-package')}
+ npm uninstall your-package'
+ Then, instead of installing the package to the root "./package.json":
+ npm install your-package
+ Install the package to "./release/app/package.json"
+ cd ./release/app && npm install your-package')}
  Read more about native dependencies at:
-${chalk.bold(
-  'https://electron-react-boilerplate.js.org/docs/adding-dependencies/#module-structure'
-)}
+ https://electron-react-boilerplate.js.org/docs/adding-dependencies/#module-structure
  `);
       process.exit(1);
     }
